requirements.txt
Flask==3.1.2
psycopg2-binary==2.9.11
gunicorn==23.0.0
python-dotenv==1.2.1
